// frontend/src/model/busLayoutResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { LayoutSeatDto } from './layoutSeatDto';

export interface BusLayoutResponse {
  id?: string;
  name?: string;
  busType?: string;
  totalSeats?: number;
  totalFloors?: number;
  totalRows?: number;
  totalCols?: number;
  description?: string;
  seats?: LayoutSeatDto[];
}