// frontend/src/model/layoutSeatDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */

export interface LayoutSeatDto {
  seatCode?: string;
  type?: string;
  floor?: number;
  row?: number;
  col?: number;
}